Extract hooks and commands logic
