+++ lib/ralph/irc.ex
@@ -15,7 +15,11 @@ defmodule Ralph.IRC do
 
   defmacro __using__(_opts) do
     quote do
+      use Ralph.IRC.Commands
+      use Ralph.IRC.Hooks
+
       import Ralph.IRC
+
       require Logger
 
       Module.register_attribute(__MODULE__, :context, persist: true)
@@ -43,63 +47,6 @@ defmodule Ralph.IRC do
           apply(__MODULE__, hook, [hook_ctx])
         end)
       end
-    end
-  end
-
-  ## Message Hooks
-
-  defmacro on_command(command, handler, param_names) do
-    handler_name = :"_on_command_hook__line_#{__CALLER__.line}"
-
-    quote do
-      def unquote(handler_name)(context, {prefix, cmd, params}) do
-        if cmd == unquote(command) do
-          params =
-            if is_function(unquote(param_names)) do
-              unquote(param_names).(params)
-            else
-              Enum.zip(unquote(param_names), params) |> Enum.into(%{})
-            end
-
-          context = Map.merge(context, params)
-
-          unquote(handler).(context)
-        end
-      end
-
-      add_hook(unquote(handler_name))
-    end
-  end
-
-  defmacro on_join(handler) do
-    quote do
-      on_command("JOIN", unquote(handler), [:channel])
-    end
-  end
-
-  defmacro on_part(handler) do
-    quote do
-      on_command("PART", unquote(handler), [:channel])
-    end
-  end
-
-  defmacro on_kick(handler) do
-    quote do
-      on_command("KICK", unquote(handler), [:channel, :target, :reason])
-    end
-  end
-
-  defmacro on_invite(handler) do
-    quote do
-      on_command("INVITE", unquote(handler), [:target, :channel])
-    end
-  end
-
-  defmacro on_privmsg(handler) do
-    quote do
-      on_command("PRIVMSG", unquote(handler), fn [target, message] ->
-        %{channel: target, target: target, message: message}
-      end)
     end
   end
 
@@ -208,35 +155,6 @@ defmodule Ralph.IRC do
   defmacro nick(nick) do
     quote do
       update_context(nick: unquote(nick))
-    end
-  end
-
-  ## Commands
-
-  # TODO: This should stay public. We should also pick a name that's consistent with the IRC nomenclature.
-  defmacro send_message(ctx, data) do
-    quote do
-      %{mod: mod, network: network} = unquote(ctx)
-      registry = Module.concat([mod, Registry])
-
-      Ralph.IRC.NetworkRegistry.lookup(registry, network, fn pid ->
-        Ralph.IRC.Connection.write(pid, unquote(data))
-      end)
-    end
-  end
-
-  defmacro privmsg(ctx, target, message) do
-    quote do
-      data = Ralph.IRC.Protocol.privmsg(unquote(target), unquote(message))
-      send_message(unquote(ctx), data)
-    end
-  end
-  def privmsg(ctx, message), do: privmsg(ctx, ctx.channel, message)
-
-  defmacro join(ctx, channel) do
-    quote do
-      data = Ralph.IRC.Protocol.join(unquote(channel))
-      send_message(unquote(ctx), data)
     end
   end
 end
